requirements.txt
# 金融新闻实时趋势分析可视化系统 - 统一依赖文件

# ===========================
# 核心依赖（所有模块共用）
# ===========================

# Redis客户端
redis==5.0.1

# 配置文件处理
PyYAML==6.0.1

# 日期时间处理
python-dateutil==2.8.2

# HTTP请求
requests==2.31.0

# ===========================
# Scraper 模块依赖
# ===========================

# Reddit 抓取
praw==7.7.1
prawcore==2.4.0

# RSS 解析
feedparser==6.0.11
newspaper3k==0.2.8
lxml_html_clean>=0.1.0

# NewsAPI 抓取
newsapi-python==0.2.7

# Twitter 抓取
snscrape==0.7.0.20230622

# 调度任务
schedule==1.2.0

# 数据导出
pandas>=2.0.0
pyarrow>=14.0.0

# ===========================
# Cleaner 模块依赖
# ===========================

# 自然语言处理
nltk>=3.8.0

# 文本处理（如果需要额外的NLP工具）
# spacy>=3.0.0

# ===========================
# Processor 模块依赖
# ===========================

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 情感分析（BERT模型）
torch>=2.0.0
transformers>=4.30.0

# 科学计算
scikit-learn>=1.3.0

# ===========================
# Visualization 后端依赖
# ===========================

# FastAPI Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# CORS中间件（已包含在FastAPI中）
python-multipart==0.0.6

# 异步支持
asyncio>=3.4.3

# ===========================
# 开发和测试依赖
# ===========================

# 测试框架
pytest==8.2.2
pytest-cov==5.0.0
pytest-mock==3.14.0
pytest-asyncio>=0.21.0

# 代码质量
# flake8>=6.0.0
# black>=23.0.0

# ===========================
# 可选依赖
# ===========================

# 如果需要GPU加速BERT（Processor模块）
# torch>=2.0.0+cu118  # CUDA 11.8版本

# 如果需要更多数据格式支持
# openpyxl>=3.1.0  # Excel支持
# lxml>=4.9.0      # XML解析

# ===========================
# 说明
# ===========================
# 1. 建议使用 Python 3.9-3.11
# 2. 统一使用虚拟环境：cs5481project
# 3. 安装命令：pip install -r requirements.txt
# 4. 如需GPU支持，请单独安装CUDA版本的PyTorch
